+++ docs/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 2016-
  * 1.0.1
  * replace space with non-breaking space after one-word prepositions in Rusyn
+ * fix false positives for sentence case on ordinal numbers in Slovak, Czech and Rusyn language
+ * replace capital letter at the beginning of the paragraph
 
 ## 2016-03-28
  * 1.0.0